Move generated version.go to sl package
